backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
openai
pydantic
python-dotenv
requests